backend/services/bim/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
jinja2
requests
python-dotenv
jpype1
mpxj
